backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
email-validator
python-jose[cryptography]
passlib[argon2]
argon2-cffi
sqlalchemy
psycopg[binary]
alembic

